refactoring to use uint64, wip

+++ main.go
@@ -115,7 +115,7 @@ func cliLoop() {
 
 				trans := objects.CreateTransaction(publicKey,
 					p,
-					currentStake/len(p2p.GetPublicKeys()),
+					currentStake/uint64(len(p2p.GetPublicKeys())),
 					publicKey.String()+time.Now().String(), //+strconv.Itoa(i),
 					secretKey)
 				channels.TransClientInput <- trans
@@ -128,7 +128,7 @@ func cliLoop() {
 			if currentStake == 0 {
 				continue
 			}
-			amount := rand.Intn(currentStake / 20)
+			amount := uint64(rand.Intn(int(currentStake) / 20))
 
 			for i := 0; i < 5; i++ {
 				receiverPK := pkList[rand.Intn(len(pkList))]
@@ -186,7 +186,7 @@ func autoTrans() {
 			receiverPK := pkList[rand.Intn(len(pkList))]
 			trans := objects.CreateTransaction(publicKey,
 				receiverPK,
-				rand.Intn(currentStake/50),
+				uint64(rand.Intn(int(currentStake)/50)),
 				strconv.Itoa(i),
 				secretKey)
 			channels.TransClientInput <- trans
